@@ -37,7 +37,7 @@ import (
 
 const (
 	// The max number of release branches we keep for CI/CD is 3, current and previous 2.
-	maxReleaseBranches = 3
+	maxReleaseBranches = 2
 
 	// Prefix for release branch names.
 	releaseBranchNamePrefix = "release-"
